requirements.txt
gphoto2==2.3.4
PySide6==6.4.0
PySide6-Addons==6.4.0
PySide6-Essentials==6.4.0
shiboken6==6.4.0
opencv-python==4.6.0.66
numpy==1.23.4
Pillow==9.2.0
psutil==5.9.3